Move tests inside src/
