+++ src/__tests__/logger/log.test.ts
@@ -1,4 +1,4 @@
-import { Log } from '../../src/logger'
+import { Log } from '../../logger'
 
 describe('Log', () => {
   const mockDate = new Date('2021-01-01T00:00:00.000Z')
