+++ src/__tests__/logger/logItem.test.ts
@@ -1,4 +1,4 @@
-import { LogItem } from '../../src/logger'
+import { LogItem } from '../../logger'
 
 describe('LogItem', () => {
   const mockDate = new Date('2021-01-01T00:00:00.000Z')
