+++ src/__tests__/logger/logger.test.ts
@@ -1,4 +1,4 @@
-import { Logger } from '../../src/logger'
+import { Logger } from '../../logger'
 
 describe('Logger', () => {
   beforeEach(() => {
